refactor: 기물 이동 경로 관리 책임 응집도 향상
