+++ src/main/java/controller/JanggiController.java
@@ -9,8 +9,8 @@ import model.JanggiGame;
 import model.Team;
 import model.board.Board;
 import model.board.BoardFactory;
-import model.board.Position;
 import model.board.TeamFormation;
+import model.coordinate.Position;
 import model.piece.Piece;
 import view.InputView;
 import view.OutputView;
