+++ src/main/java/model/board/Board.java
@@ -4,6 +4,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import model.coordinate.Position;
 import model.piece.Piece;
 
 public class Board {
