+++ src/main/java/model/board/TeamFormation.java
@@ -2,6 +2,7 @@ package model.board;
 
 import java.util.Map;
 import model.Team;
+import model.coordinate.Position;
 import model.piece.Piece;
 
 public record TeamFormation(Team team, FormationType type) {
