--- src/main/java/model/movement/MoveStrategy.java
@@ -1,9 +1,0 @@
-package model.movement;
-
-import java.util.List;
-import model.board.Position;
-
-public interface MoveStrategy {
-
-    List<Position> extractPath(Position start, Position end);
-}
