--- src/main/java/model/movement/OneStepStrategy.java
@@ -1,16 +1,0 @@
-package model.movement;
-
-import java.util.List;
-import model.board.Position;
-
-public class OneStepStrategy implements MoveStrategy {
-
-    @Override
-    public List<Position> extractPath(Position start, Position end) {
-        Displacement displacement = end.toDisplacement(start);
-        Direction cardinal = displacement.extractCardinal();
-
-        Position step = cardinal.move(start);
-        return List.of(step);
-    }
-}
