+++ src/main/java/model/piece/Guard.java
@@ -1,7 +1,8 @@
 package model.piece;
 
+import java.util.List;
 import model.Team;
-import model.board.Position;
+import model.coordinate.Position;
 
 public class Guard extends Piece {
 
@@ -11,6 +12,11 @@ public class Guard extends Piece {
 
     @Override
     protected void validateMove(Position current, Position next) {
+        throw new IllegalArgumentException("1단계 궁성 영역 미구현");
+    }
+
+    @Override
+    protected List<Position> extractPath(Position current, Position next) {
         throw new IllegalArgumentException("1단계 궁성 영역 미구현");
     }
 }
