+++ src/main/java/model/piece/Soldier.java
@@ -1,8 +1,9 @@
 package model.piece;
 
+import java.util.List;
 import model.Team;
-import model.board.Position;
-import model.movement.Displacement;
+import model.coordinate.Displacement;
+import model.coordinate.Position;
 
 public class Soldier extends Piece {
 
@@ -26,5 +27,10 @@ public class Soldier extends Piece {
         if (!(displacement.isForwardBy(forwardDirection) || displacement.isSideOneStep())) {
             throw new IllegalArgumentException("졸이 이동할 수 없는 위치입니다.");
         }
+    }
+
+    @Override
+    protected List<Position> extractPath(Position current, Position next) {
+        return List.of();
     }
 }
